Refactor to allow shiftDates to be used by other plugins
