+++ jgclark.EventHelpers/plugin.json
@@ -8,8 +8,8 @@
   "plugin.author": "jgclark",
   "plugin.url": "https://github.com/NotePlan/plugins/tree/main/jgclark.EventHelpers",
   "plugin.changelog": "https://github.com/NotePlan/plugins/tree/main/jgclark.EventHelpers/CHANGELOG.md",
-  "plugin.version": "0.22.1",
-  "plugin.lastUpdateInfo": "0.22.1: - extend /shift dates to cover weekly, monthly, quarterly and yearly dates as well as daily ones.\n0.22.0: can now use events() template calls in Weekly notes.\n0.21.3: bug fix adding time blocks to calendar.\n0.21.2: /shiftDates now covers more cases.\n0.21.1: add 'Yes to all' to option to create time blocks. Extend '/shift dates' to work on week dates.\n0.21.0: improvements to '/shift dates' and '/process date offsets'.",
+  "plugin.version": "0.22.2",
+  "plugin.lastUpdateInfo": "0.22.2: ???\n0.22.1: extend /shift dates to cover weekly, monthly, quarterly and yearly dates as well as daily ones.\n0.22.0: can now use events() template calls in Weekly notes.\n0.21.3: bug fix adding time blocks to calendar.\n0.21.2: /shiftDates now covers more cases.\n0.21.1: add 'Yes to all' to option to create time blocks. Extend '/shift dates' to work on week dates.\n0.21.0: improvements to '/shift dates' and '/process date offsets'.",
   "plugin.dependencies": [],
   "plugin.script": "script.js",
   "plugin.isRemote": "false",
@@ -40,10 +40,21 @@
     {
       "name": "shift dates",
       "alias": [
-        "offset"
+        "offset",
+        "shift"
       ],
       "description": "takes dates in the selection and shifts them forwards or backwards by a given date interval",
       "jsFunction": "shiftDates"
+    },
+    {
+      "name": "shiftDatesCore",
+      "description": "Entry point for shiftDates for other commands",
+      "jsFunction": "shiftDatesCore",
+      "parameters": [
+        "note",
+        "parasToProcess",
+        "interval"
+      ]
     },
     {
       "name": "listDaysEvents",
